more OOP (almost done)
