+++ data/duke.txt
@@ -1,2 +1,3 @@
 T,1,eat noodles
 D,0,do homework,by,2019-12-01
+T,0,eat food
